fix: transfer amount uint160

+++ src/SablierV2ProxyTarget.sol
@@ -165,7 +165,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
@@ -220,7 +220,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
@@ -391,7 +391,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
@@ -446,7 +446,7 @@ contract SablierV2ProxyTarget is
         // Calculate the sum of all of stream amounts. It is safe to use unchecked addition because one of the create
         // transactions will revert if there is overflow.
         uint256 i;
-        uint128 transferAmount;
+        uint160 transferAmount;
         for (i = 0; i < batchSize;) {
             unchecked {
                 transferAmount += batch[i].totalAmount;
